% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_plugins.R
\name{get_vamp_plugin_dir}
\alias{get_vamp_plugin_dir}
\title{Get the appropriate Vamp plugin directory for the current OS}
\usage{
get_vamp_plugin_dir(user_dir = TRUE)
}
\arguments{
\item{user_dir}{If TRUE, use user-specific directory; if FALSE, use system directory}
}
\value{
Path to the Vamp plugin directory
}
\description{
Get the appropriate Vamp plugin directory for the current OS
}
\note{
On Windows, the system directory (C:\\Program Files\\Vamp Plugins) may require administrator rights.
User directory is recommended for non-admin installations.
}
